base,progress: move ProgressParams to base.go
