+++ progress_gojay.go
@@ -11,70 +11,6 @@ import (
 	"github.com/francoispqt/gojay"
 )
 
-// ProgressToken is the progress token provided by the client or server.
-//
-// @since 3.15.0.
-type ProgressToken interface{}
-
-// NewProgressToken returns a new ProgressToken.
-//nolint:gocritic
-func NewProgressToken(s string) *ProgressToken {
-	var iface interface{} = s
-	return (*ProgressToken)(&iface)
-}
-
-// NewNumberProgressToken returns a new number ProgressToken.
-//nolint:gocritic
-func NewNumberProgressToken(n int32) *ProgressToken {
-	var iface interface{} = n
-	return (*ProgressToken)(&iface)
-}
-
-//nolint:gocritic
-func encodeProgressToken(enc *gojay.Encoder, key string, v *ProgressToken) {
-	if v == nil {
-		return
-	}
-	enc.AddInterfaceKey(key, (interface{})(*v))
-}
-
-//nolint:gocritic
-func decodeProgressToken(dec *gojay.Decoder, k, key string, v *ProgressToken) error {
-	if v == nil || k != key {
-		return nil
-	}
-	return dec.Interface((*interface{})(v))
-}
-
-// MarshalJSONObject implements gojay.MarshalerJSONObject.
-func (v *ProgressParams) MarshalJSONObject(enc *gojay.Encoder) {
-	encodeProgressToken(enc, keyToken, &v.Token)
-	enc.AddInterfaceKey(keyValue, v.Value)
-}
-
-// IsNil implements gojay.MarshalerJSONObject.
-func (v *ProgressParams) IsNil() bool { return v == nil }
-
-// UnmarshalJSONObject implements gojay.UnmarshalerJSONObject.
-func (v *ProgressParams) UnmarshalJSONObject(dec *gojay.Decoder, k string) error {
-	switch k {
-	case keyToken:
-		return decodeProgressToken(dec, k, keyToken, &v.Token)
-	case keyValue:
-		return dec.Interface(&v.Value)
-	}
-	return nil
-}
-
-// NKeys implements gojay.UnmarshalerJSONObject.
-func (v *ProgressParams) NKeys() int { return 2 }
-
-// compile time check whether the ProgressParams implements a gojay.MarshalerJSONObject and gojay.UnmarshalerJSONObject interfaces.
-var (
-	_ gojay.MarshalerJSONObject   = (*ProgressParams)(nil)
-	_ gojay.UnmarshalerJSONObject = (*ProgressParams)(nil)
-)
-
 // MarshalJSONObject implements gojay.MarshalerJSONObject.
 func (v *WorkDoneProgressBegin) MarshalJSONObject(enc *gojay.Encoder) {
 	enc.StringKey(keyKind, string(v.Kind))
