+++ progress_test.go
@@ -6,101 +6,10 @@ package protocol
 
 import (
 	"fmt"
-	"reflect"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
 )
-
-func testProgressParams(t *testing.T, marshal marshalFunc, unmarshal unmarshalFunc) {
-	const wantWorkDoneToken = "156edea9-9d8d-422f-b7ee-81a84594afbb"
-	const want = `{"token":"` + wantWorkDoneToken + `","value":"testValue"}`
-
-	token := NewProgressToken(wantWorkDoneToken)
-	wantType := ProgressParams{
-		Token: *token,
-		Value: "testValue",
-	}
-
-	t.Run("Marshal", func(t *testing.T) {
-		t.Parallel()
-
-		tests := []struct {
-			name           string
-			field          ProgressParams
-			want           string
-			wantMarshalErr bool
-			wantErr        bool
-		}{
-			{
-				name:           "Valid",
-				field:          wantType,
-				want:           want,
-				wantMarshalErr: false,
-				wantErr:        false,
-			},
-		}
-
-		for _, tt := range tests {
-			tt := tt
-			t.Run(tt.name, func(t *testing.T) {
-				t.Parallel()
-
-				got, err := marshal(&tt.field)
-				if (err != nil) != tt.wantMarshalErr {
-					t.Fatal(err)
-				}
-
-				if diff := cmp.Diff(string(got), tt.want); (diff != "") != tt.wantErr {
-					t.Errorf("%s: wantErr: %t\n(-got, +want)\n%s", tt.name, tt.wantErr, diff)
-				}
-			})
-		}
-	})
-
-	t.Run("Unmarshal", func(t *testing.T) {
-		t.Parallel()
-
-		tests := []struct {
-			name             string
-			field            string
-			want             ProgressParams
-			wantUnmarshalErr bool
-			wantErr          bool
-		}{
-			{
-				name:             "Valid",
-				field:            want,
-				want:             wantType,
-				wantUnmarshalErr: false,
-				wantErr:          false,
-			},
-		}
-
-		for _, tt := range tests {
-			tt := tt
-			t.Run(tt.name, func(t *testing.T) {
-				t.Parallel()
-
-				var got ProgressParams
-				if err := unmarshal([]byte(tt.field), &got); (err != nil) != tt.wantUnmarshalErr {
-					t.Fatal(err)
-				}
-
-				if diff := cmp.Diff(got, tt.want, cmpopts.IgnoreFields(ProgressParams{}, "Token")); (diff != "") != tt.wantErr {
-					t.Errorf("%s: wantErr: %t\n(-got, +want)\n%s", tt.name, tt.wantErr, diff)
-				}
-
-				if token := got.Token; !reflect.ValueOf(token).IsZero() {
-					if diff := cmp.Diff(fmt.Sprint(token), wantWorkDoneToken); (diff != "") != tt.wantErr {
-						t.Errorf("%s: wantErr: %t\n(-got, +want)\n%s", tt.name, tt.wantErr, diff)
-					}
-				}
-			})
-		}
-	})
-}
 
 func testWorkDoneProgressBegin(t *testing.T, marshal marshalFunc, unmarshal unmarshalFunc) {
 	const (
